Remove Saved Props From Most Fields
Remove most of the saved props from most of the fields, with the expect
of single_select_typeahead. In order to test if the saved list is being
used without having to use props to active saved label for that field.

This will also test if the user can force the saved label to display for
a single field as well.