Fix Fieldset Fields To Show Saved On Correct Field
Fix the Fieldset Fields to show saved label, on the correct fields.
Since correctly the filter is removing an elements, in the Fieldset
Field array, that is not modified. Because of this the saved label is
only displaying, on the first field, as that will always be only element
in the array.