Fix Running Coveralls On Fork Pull Request
Fix running of Coveralls if the pull request is coming from a fork
version, of the repo. Because CricleCI blocks custom environment
variables from running, if the pull request is coming, from a different
repo. This is to prevent a developer from scrapping environment
variables when the application runs on CricleCI.